+++ client/app/components/particles/Emitter.tsx
