+++ client/app/components/particles/LinearSpline.tsx
