+++ client/app/components/particles/SpriteCell.tsx
